+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jala-banyu/button
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.5.4
+  - @jala-banyu/ripple@5.0.4
+  - @jala-banyu/spinner@5.0.4
+
 ## 5.0.3
 
 ### Patch Changes