+++ packages/components/image/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jala-banyu/image
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.5.4
+
 ## 5.0.3
 
 ### Patch Changes