+++ packages/components/listbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/listbox
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.5.4
+  - @jala-banyu/input@5.0.4
+
 ## 2.0.3
 
 ### Patch Changes