+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/menu
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.5.4
+  - @jala-banyu/checkbox@2.0.4
+
 ## 5.0.3
 
 ### Patch Changes