+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jala-banyu/switch
 
+## 5.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.5.4
+
 ## 5.0.4
 
 ### Patch Changes